tools/bigquery-tagging/requirements.txt
google-cloud-bigquery==1.9.0
Flask==1.0.2